Architectural changes, quite a few revisions
Re-oriented the system around a single Service Centre that lends out global access to different services (e.g. logging, memory management, rendering, physics...) as appropriate; edited out the more (seemingly) inelegant parts of the tutorial code to provide a sleaker, more effective, and more reasonable system.

Created stubs for memory management and logging (to be tackled next weekend), also implemented a Map template class for storing associations between arbitrary types (used by the ServiceCentre) as well as a KnownLengthArray (not implemented, but useful for variable-length, heavily-iterated arrays). 

Renamed the "Engine" class to the more appropriate "Application" and added in headers for constants (hey, if you're going to have a bunch of contextless #defines you might as well store them in one place) and typedefs (mainly so that signedness and exact length can be stored in various typenames instead of implied). Also added in a friend's CRT memory tracker so that leaks can be found and sealed without trekking through every file in the program :)