Implemented memory management, edited service centre
Memory management works through the Stack Allocator described in Game Engine Architecture. Stack Allocators are powerful and fairly easy to write, but inelegant for de-allocations; you have to call the de-allocator, then manually call the relevant destructors and send the relevant pointers to the null address. Destruction/banishment would be simplified with smart pointers, but the basic design of the Stacl Allocator (stack-based, as you might have guessed) would still run against the random-access paradigm used by C++ and ultimately create awkward timing issues where frees can only occur if there's a nearby marker.

For all those reasons, I'm OK to continue with the stack allocator while Athru is relatively simple but less certain about using it long-term. We look at an alternative model (known as the Memory Pool) which _does_ work well with random allocations/de-allocations, sometime in Term 1 this year (2017) at AIE and once I've created a working one of those I wouldn't mind embedding it into this project.

That said, Game Engine Architecture does contain a description of a Memory Pool and how to manage fragmentation (one of the downsides of using the pool approach; memory can fragment), so if we make faster progress than expected I'd be happy to pull the Stack Allocator out of the project and build the system around a Memory Pool instead; at the moment, though, the system itself isn't complicated enough to justify a complex memory model.

Earlier versions of the Service Centre used an associative array with a string lookup and a single register/access function; this was easy to use, but very slow. The service centre's been edited to have distinct static register/access functions for every service (as described in Game Programming Patterns), which is quite awkward to use but very fast. Since we aren't going to be writing that many distinct services in the end (maybe 5, 10 more at most?) I figure the trade-off is worth it.

Some parts of the system could be getting a little dusty, and I wouldn't mind a refactor some time in the future to improve all the bits we haven't noticed yet. I'd favour getting a basic system done first though, so any big clean-ups should probably be left until after we can render out a basic scene.

Major unimplemented features are (listed by priority):
- Vector/Matrix library
- Render pipeline and GUI
- Sound
- Physics

Once everything in that list is implemented, we can start building an actual game :D. We'll still need to do gameplay stuffs, all the procedural systems, AI, and other things like that, but the minimum to build something that you can explore and interact with will be there :).