Corrected vertex rendering, implemented vertex transparency
So, after lots of digging I found two niggly little mistakes that were probably hiding the triangle:

- The vectors fed into the view matrix were laid out left-to-right, when (because XMVECTORs are built around SSE) the correct order is right-to-left

- The position vectors within each [Vertex] were passing into the shader with RGB (3-part vector) format, even though they were specified within the shader as RGBA (4-part vector). This ended up displacing the vertices as they passed into the shader and simultaneously (a) masking off the [w] part of each vertex position, and (b) shifting each color one column to the right. 

Both of those have been fixed (and I've reverted the alignment hack I used when I _thought_ I'd fixed the color issues), and now the triangle displays with the expected colors and with each vertex sitting at (or very close to) the expected coordinates. 

It's unrelated, but I also added vertex transparency. It's a nice thing to have around, and the solution (set up a blend state, create the blend state, pass the blend state into the device context) was short and sharp enough that I thought I might as well slip it in :P. It took a few minutes to debug once I'd written the code itself, but it was still fairly simple to implement for such a potentially-useful feature.