Readability + aesthetic improvements
Added a custom constructor to the [Vertex] struct to make it easier to use and more readable during array initialisation (e.g while making the verts for a Boxecule).

(lots of progress between updates! I've also replaced the triangle from the last commit with a nice, volumetric, only slightly broken lozenge-style cubey thing :P. It's much closer to a full, scripted "boxecule" than the triangle was, so I've renamed it to that and started gradually changing it back into the sort of cool atomic voxel class I was going for before everything broke and I had to nuke everything in the BIG_FREAKIN_PATCH to fix it).

Also edited the colors in the test render a little bit to make each vertex more distinct, and rotated the Boxecule to face the camera at an angle to prettify it and double-check that everything was working...and basically only the rear view (facing the screen and parallel to the camera) worked, I just had backface culling off in the debugger. It looks better now, but there's also a butt-ton of z-fighting from redundant tris; it's definitely something I'm going to have to come back to tomorrow/tonight.