Mostly implemented frustum culling
Wrote most of the process + a frustum class, just need to implement construction, test, and debug :)
NB: The "pointer weirdness" in the last comit was me overthinking how to access elements from a heaped one-dimensional array of pointers; since it was causing memory to allocate a few indices off from where it was supposed to go, nothing displayed and the renderer would break when it reached an uninitialised Boxecule/a Boxecule initialised with garbage. Fixing the pointer weirdness and undoing some paranoid "fixes" I made during the debug process got the game to run on the first and second tries, but then it stopped working again because (I think) I was feeding too much data into the GPU (385875 voxels drawn every frame...yeah). I was always going to add frustum culling, but I'm focussing on it slightly earlier than was going to so I can boost my framerate during testing  + reliably see some output when I start debugging :P