Fixed boxecule generation, edited DirectX handler class
More-or-less forgot about frustum generation, so I'll have to implement that tonight. I thought I was killing the GPU by making too many voxels, but it turns out the Boxecules (and the chunks >.<) were on the stack and destructed as soon as they left the scope of the constructor where they were generated. Moving them to the heap and overloading the [new] + [delete] operators to use Athru's memory allocator helped with that.

There were a few parameters in the DirectX constructor that were global anyway, so I removed those and edited all the constructor's references.

Even if I didn't write any code to construct the frustum planes, I FINALLY got the math together to build them from screen near/far and the camera field-of-view (thanks Unity docs :P). That should make them fairly easy to build once I get around to it :D