Minor bugfixes, lots of research into remaining culling issues
Reverse-culling effect was produced because I was storing the frustum data backwards; that's fixed, but now culling begins cutting across the scene as you travel further along the Z-axis. It feels basically unfixable at the moment, so I'm probably going to ask about it on StackOverflow and see what they say.

It's not a /severe/ bug, just annoying; I think I'll go through the rest of our tasklist in the meantime.