Big rendering + scene update
Everything is rendered with raymarching, and all objects are represented with specific versions of different distance functions :D

Raymarching can be effectively scaled up to full-on path tracing (= highly realistic lighting), and rendering with ray-marched distance functions allows Athru to display extremely smooth and continuous figures that would be difficult to realize with polygons or voxel geometry.

Distance functions are also inherently specific to the area defined by their scaling factor, making it trivial to plot every function in a scene inside a single compute-shading pass.